+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.5.4]
+
+## [v0.6.0]
+## Changed
+- Sites with no coverage now report "nan" methylation frequency and score.
 ## Added
+- Option `--aggregate` to pair information from two strands and output additional files.
 - Support for ChEBI codes in C and Python pileup API.
 
 ## [v0.5.3]